requirements.txt
Flask==3.0.3
keras==2.10.0
Pillow==10.4.0
numpy==1.26.4
tensorflow==2.10.0